feat(api,web): migrate portion resource to plural routes and path params
