--- apps/web/src/routes/portion/index.tsx
@@ -1,9 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/portion/")({
-  component: RouteComponent,
-});
-
-function RouteComponent() {
-  return <div>Hello "/portion/"!</div>;
-}
